Move views into dedicated package within `metrics/public_api/views/` split by base, root and linked views
