+++ metrics/public_api/views/__init__.py
@@ -1,0 +1,14 @@
+from .linked_views import (
+    GeographyDetailView,
+    GeographyListView,
+    GeographyTypeDetailView,
+    GeographyTypeListView,
+    MetricListView,
+    SubThemeDetailView,
+    SubThemeListView,
+    ThemeDetailView,
+    ThemeListView,
+    TopicDetailView,
+    TopicListView,
+)
+from .root_view import PublicAPIRootView
